Implement _Media base class for Image and Video
